Restructure simulations into modular library and adapters
