+++ .vscode/README.md
@@ -1,6 +1,6 @@
 This folder stores VS Code configuration files for launching and working with the simulation.
 The provided `launch.json` lets you run `run_simulation.py` directly from the editor.
 It also includes configurations `Blender Deck Simulation` and `Blender Hull Simulation` which start Blender
-via their respective `blender_starter.py` scripts (requires the environment variable `BLENDER_PATH`).
+via their respective `starter.py` scripts (requires the environment variable `BLENDER_PATH`).
 `settings.json` contains defaults for previewing generated results.
