+++ simulations/README.md
@@ -2,8 +2,8 @@
 
 This folder hosts example scripts and data used to generate visualizations for the Sphere Station project.
 
-The `scripts` subdirectory contains the Python module `deck_calculations_script.py`, which builds geometry for the rotating sphere station and renders animations.
+The `deck_calculator` subdirectory exposes `deck_calculations_script.py` as a lightweight wrapper around the reusable library.
 
-`blender_simulation` and `blender_hull_simulation` provide example Blender scripts for visualizing the decks and the outer hull respectively.
+`blender_deck_simulator` and `blender_hull_simulation` provide example Blender scripts for visualising the decks and the outer hull respectively.
 
 Simulation code is provided under the MIT license for educational purposes. Refer to `LICENSE-MIT` for details. Other content within this repository may be subject to the proprietary license described in the root `README.md`.
