+++ simulations/blender_deck_simulator/deck_3d_construction_data.csv
