+++ simulations/blender_hull_simulation/README.md
@@ -2,5 +2,5 @@
 
 This folder contains a minimal Blender Python script that generates a simple hull model for the Sphere Station. It mirrors the approach of the deck simulation but focuses solely on the outer hull.
 
-* **blender_hull_simulator.py** – Creates a basic hull based on `deck_3d_construction_data.csv`.
-* **blender_starter.py** – Convenience launcher that starts Blender using the environment variable `BLENDER_PATH`.
+* **adapter.py** – Creates a basic hull based on `deck_3d_construction_data.csv`.
+* **starter.py** – Convenience launcher that starts Blender using the environment variable `BLENDER_PATH`.
