+++ simulations/blender_station_simulator/README.md
@@ -1,0 +1,2 @@
+# Blender Station Simulator
+Run this script to start the full station simulation.
