+++ simulations/deck_calculator/deck_calculations_script.py
@@ -1,0 +1,5 @@
+"""Compatibility wrapper for the :mod:`sphere_space_station_simulations` package."""
+
+from simulations.sphere_space_station_simulations import SphereDeckCalculator
+
+__all__ = ["SphereDeckCalculator"]
