+++ simulations/deck_calculator/starter.py
@@ -1,7 +1,12 @@
-from simulations.library import SphereDeckCalculator
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
+
+from simulations.deck_calculator.adapter import SphereDeckCalculator
 
 
-def main():
+def main() -> None:
     calculator = SphereDeckCalculator(
         "Deck Dimensions of a Sphere",
         sphere_diameter=127.0,
@@ -14,8 +19,18 @@ def main():
     )
 
     calculator.calculate_dynamics_of_a_sphere(angular_velocity=0.5)
+    print(calculator.to_string())
     calculator.to_csv("deck_dimensions.csv")
     calculator.to_html("deck_dimensions.html")
+    calculator.to_3D_animation_show_all_decks("deck_animation.html")
+    calculator.to_3D_animation_rotate_all_decks("deck_animation_rotate.html")
+    calculator.to_3D_animation_rotate_hull_with_windows(
+        "hull_with_windows_animation.html",
+        frames=25,
+        frames_per_second=5,
+        rotation_axis="Z",
+    )
+    calculator.to_3D_animation_rotate_hull("hull_animation.html")
     calculator.to_3D_animation_rotate_hull_with_gravity_zones(
         "hull_with_gravity_zones_animation.html",
         frames=25,
