--- simulations/library/README.md
@@ -1,1 +1,0 @@
-# Library\nThis package contains reusable simulation components for the Sphere Station.
