+++ simulations/sphere_space_station_simulations/README.md
@@ -1,0 +1,3 @@
+# Sphere Space Station Simulations Library
+
+This package bundles the core calculations, animation helpers and data export functionality used by the example simulations.  The modules here can be reused by the thin adapters located in the individual simulation folders.
