--- simulations/start_simulation/README.md
@@ -1,2 +1,0 @@
-# Simulation Entry Point
-Run this script to start the station simulation.
