+++ simulations/tests/test_geometry.py
@@ -2,7 +2,7 @@ import sys
 import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-from simulations.scripts.deck_calculations_script import SphereDeckCalculator
+from simulations.sphere_space_station_simulations import SphereDeckCalculator
 
 
 def test_window_count():
